fix syntax errors

+++ Lab 4/src/Account.java
@@ -38,7 +38,7 @@ class Account
 
 	/* Returns the balance of the Account
 	 */
-	public int getBalance()
+	public double getBalance()
 	{
 		return balance;
 	}
